Refactor(2주차-미션1): Props 타입 별도 파일로 분리
