+++ 윤-정윤주/2주차/2nd-mission1/src/components/TodoList/TodoList.tsx
@@ -1,12 +1,4 @@
-import type { TTodo } from "../types/todo";
-
-interface TodoListProps {
-    title: string;
-    todos: TTodo[];
-    buttonLabel: string;
-    buttonColor: string;
-    onClick: (todo: TTodo) => void;
-}
+import type { TodoListProps } from "./TodoList.types";
 
 const TodoList = ({
     title,
