+++ 윤-정윤주/2주차/2nd-mission1/src/components/TodoList/TodoList.types.ts
@@ -1,0 +1,9 @@
+import type { TTodo } from "../../types/todo";
+
+export interface TodoListProps {
+    title: string;
+    todos: TTodo[];
+    buttonLabel: string;
+    buttonColor: string;
+    onClick: (todo: TTodo) => void;
+}
